超管管理的课表管理1.0 暂时使用默认模板，下次迭代使用现有模板

+++ src/main/java/org/example/pingpongsystem/entity/SchoolEntity.java
@@ -25,7 +25,7 @@ public class SchoolEntity {
 
     private String email;
 
-    private int adminId;
+    private Long adminId;
 
     @Column(nullable = false)
     private int table_num;
